reorganization of code
